requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.3
langgraph==0.2.26
pydantic==2.9.2
tiktoken==0.7.0
rich==13.9.3
ipython==8.28.0
beautifulsoup4==4.12.3
langchain-groq
python-dotenv
